+++ src/main/java/org/akcjamis/webapp/service/ChristmasPackageService.java
@@ -77,7 +77,7 @@ public class ChristmasPackageService {
      *  @return the list of entities
      */
     @Transactional(readOnly = true)
-    public Page<ChristmasPackage> getList(Long eventId, Pageable pageable){
+    public Page<ChristmasPackage> getList(Short eventId, Pageable pageable){
         log.debug("Request to get all ChristmasPackages");
         return christmasPackageRepository.getList(eventId, pageable);
     }
